requirements.txt
python-telegram-bot[rate-limiter]
pydub
toml
yt-dlp
coloredlogs